pyproject.toml
[tool.poetry]
name = "analyzer-banking-operations"
version = "0.1.0"
description = ""
authors = ["Egor Arhipov <111359850+Maurdihen@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.2.2"
flake8 = "^7.1.0"
pandas = "^2.2.2"
mypy = "^1.10.1"
isort = "^5.13.2"
black = "^24.4.2"
openpyxl = "3.1.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
yfinance = "^0.2.40"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
ignore_missing_imports = true
exclude = 'venv'

[tool.isort]
line_length = 120
profile = "black"

multi_line_output = 3

[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
)
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
